+++ school-bus/guns-api/src/main/java/com/stylefeng/guns/rest/bus/dto/PageBusRequest.java
@@ -14,7 +14,7 @@ import lombok.Data;
 public class PageBusRequest extends AbstractRequest {
 
     // 当前页
-    private Integer currentPage;
+    private Long currentPage;
     // 每页数量
-    private Integer pageSize;
+    private Long pageSize;
 }
