+++ school-bus/guns-api/src/main/java/com/stylefeng/guns/rest/order/dto/EvaluateRequest.java
@@ -12,11 +12,11 @@ import lombok.Data;
 
 @Data
 public class EvaluateRequest extends AbstractRequest {
-    private Integer userId;
+    private Long userId;
     // 评价状态
     private String evaluateStatus;
     // 当前页
-    private Integer currentPage;
+    private Long currentPage;
     // 每页数量
-    private Integer pageSize;
+    private Long pageSize;
 }
