+++ school-bus/guns-gateway/src/main/java/com/stylefeng/guns/rest/modular/form/CountPageInfo.java
@@ -15,11 +15,11 @@ import lombok.Data;
 @ApiModel(value = "分页实体", description = "分页请求信息")
 public class CountPageInfo {
 
-    @ApiModelProperty(name = "currentPage", value = "当前页", example = "1", dataType = "int", required = true)
-    private Integer currentPage;
+    @ApiModelProperty(name = "currentPage", value = "当前页", example = "1", dataType = "long", required = true)
+    private Long currentPage;
 
-    @ApiModelProperty(name = "pageSize", value = "每页的条目数量", example = "4", dataType = "int", required = true)
-    private Integer pageSize;
+    @ApiModelProperty(name = "pageSize", value = "每页的条目数量", example = "4", dataType = "long", required = true)
+    private Long pageSize;
 
     @ApiModelProperty(name = "busStatus", value = "0:沙河->清水河，1:清水河->沙河", example = "0", dataType = "String", required = true)
     private String busStatus;
