+++ school-bus/guns-gateway/src/main/java/com/stylefeng/guns/rest/modular/form/OrderPageInfo.java
@@ -14,9 +14,9 @@ import lombok.Data;
 @Data
 @ApiModel(value = "分页实体", description = "分页请求信息")
 public class OrderPageInfo {
-    @ApiModelProperty(name = "currentPage", value = "当前页", example = "1", dataType = "int", required = true)
-    private Integer currentPage;
+    @ApiModelProperty(name = "currentPage", value = "当前页", example = "1", dataType = "long", required = true)
+    private Long currentPage;
 
-    @ApiModelProperty(name = "pageSize", value = "每页的条目数量", example = "4", dataType = "int", required = true)
-    private Integer pageSize;
+    @ApiModelProperty(name = "pageSize", value = "每页的条目数量", example = "4", dataType = "long", required = true)
+    private Long pageSize;
 }
